Datatype da PK idSessao definido como int. Corrige o problema de interpretação de dados quando um item de pesquisa era selecionado.

+++ trunk/biblioteca-eeep/src/fvsosp/sessao/SessaoDAO.java
@@ -44,7 +44,7 @@ public class SessaoDAO extends GenericDAO<Sessao> {
 
     }
     
-    public Sessao pesquisarCodigo(short codigo) {
+    public Sessao pesquisarCodigo(int codigo) {
         Sessao sessao = null;
         try {
             this.setSessao(HibernateUtil.getSessionFactory().openSession());
